feat: add remaining SEO things
